+++ R/wflow_git_push.R
@@ -151,6 +151,20 @@ wflow_git_push <- function(remote = NULL, branch = NULL,
   # Push! ----------------------------------------------------------------------
 
   if (!dry_run) {
+    # First check for and execute any pre-push hooks. libgit2 does not support
+    # this.
+    pre_push_file <- file.path(git2r_workdir(r), ".git/hooks/pre-push")
+    pre_push_file_rel <- fs::path_rel(pre_push_file, start = getwd())
+    if (fs::file_exists(pre_push_file)) {
+      message(glue::glue("Executing pre-push hook in {pre_push_file_rel}"))
+      hook <- suppressWarnings(system(pre_push_file, intern = TRUE))
+      message(hook)
+      if (attributes(hook)$status != 0) {
+        stop(glue::glue("Execution stopped by {pre_push_file_rel}"),
+             call. = FALSE)
+      }
+    }
+
     tryCatch(git2r::push(r, name = remote,
                          refspec = paste0("refs/heads/", branch),
                          force = force, credentials = credentials),