+++ R/wflow_start.R
@@ -157,6 +157,16 @@
 #' @param change_wd logical (default: \code{TRUE}). Change the working
 #'   directory to the \code{directory}.
 #'
+#' @param disable_remote logical (default: \code{FALSE}). Create a Git
+#'   \href{https://git-scm.com/book/en/v2/Customizing-Git-Git-Hooks}{pre-push
+#'   hook} that prevents pushing to a remote Git repository (i.e. using
+#'   \code{\link{wflow_git_push}}). This is useful for extremely confidential
+#'   projects that cannot be shared via an online Git hosting service (e.g.
+#'   GitHub or GitLab). The hook is saved in the file
+#'   \code{.git/hooks/pre-push}. If you change your mind and want to push the
+#'   repository, you can delete that file. Note that this option is only
+#'   available if \code{git = TRUE}.
+#'
 #' @param dry_run logical (default: \code{FALSE}). When \code{dry_run
 #'   = TRUE}, the actions are previewed without executing them.
 #'
@@ -192,6 +202,8 @@
 #'    \item{overwrite}{The input argument \code{overwrite}.}
 #'
 #'    \item{change_wd}{The input argument \code{change_wd}.}
+#'
+#'    \item{disable_remote}{The input argument \code{disable_remote}.}
 #'
 #'    \item{dry_run}{The input argument \code{dry_run}.}
 #'
@@ -232,6 +244,7 @@ wflow_start <- function(directory,
                         existing = FALSE,
                         overwrite = FALSE,
                         change_wd = TRUE,
+                        disable_remote = FALSE,
                         dry_run = FALSE,
                         user.name = NULL,
                         user.email = NULL) {
@@ -247,6 +260,8 @@ wflow_start <- function(directory,
     stop("overwrite must be a one element logical vector: ", overwrite)
   if (!is.logical(change_wd) | length(change_wd) != 1)
     stop("change_wd must be a one element logical vector: ", change_wd)
+  if (!is.logical(disable_remote) | length(disable_remote) != 1)
+    stop("disable_remote must be a one element logical vector: ", disable_remote)
   if (!is.logical(dry_run) | length(dry_run) != 1)
     stop("dry_run must be a one element logical vector: ", dry_run)
   if (!(is.null(user.name) | (is.character(user.name) | length(user.name) != 1)))
@@ -280,6 +295,11 @@ wflow_start <- function(directory,
   # Require that user.name and user.email be set locally or globally
   if (git && is.null(user.name) && is.null(user.email)) {
     check_git_config(path = directory, "`wflow_start` with `git = TRUE`")
+  }
+
+  # Do not allow git=FALSE and disable_remote = TRUE
+  if (!git && disable_remote) {
+    stop("disable_remote is only available if git=TRUE")
   }
 
   # Create directory if it doesn't already exist
@@ -358,6 +378,17 @@ wflow_start <- function(directory,
     } else{
       commit <- git2r::commit(repo, message = "Start workflowr project.")
     }
+    # Create pre-push hook to prevent pushing confidential projects
+    if (disable_remote) {
+      pre_push_file <- file.path(git2r_workdir(repo), ".git/hooks/pre-push")
+      if (!fs::file_exists(pre_push_file) || overwrite) {
+        # extras is a list defined in infrastructure.R
+        cat(glue::glue(extras[["disable_remote"]]), file = pre_push_file)
+      }
+      if (!file_is_executable(pre_push_file)) {
+        fs::file_chmod(pre_push_file, "a+x")
+      }
+    }
   }
 
   # Prepare output -------------------------------------------------------------
@@ -368,6 +399,7 @@ wflow_start <- function(directory,
             existing = existing,
             overwrite = overwrite,
             change_wd = change_wd,
+            disable_remote = disable_remote,
             dry_run = dry_run,
             user.name = user.name,
             user.email = user.email,
@@ -405,6 +437,9 @@ print.wflow_start <- function(x, ...) {
     } else {
       cat("- Files will not be commited with Git\n")
     }
+    if (x$disable_remote) {
+      cat("- Pushing to remote repository will be disabled\n")
+    }
   } else {
     cat("wflow_start:\n")
     if (x$existing) {
@@ -437,6 +472,9 @@ print.wflow_start <- function(x, ...) {
       }
     } else {
       cat("- No Git repo\n")
+    }
+    if (x$disable_remote) {
+      cat("- Pushing to remote repository is disabled\n")
     }
   }